refactor code — create separate decorator file
