+++ cashman-flask-project/cashman/index.py
@@ -1,33 +1,47 @@
 """Python Flask RESTful API Auth0 integration example
 """
-
-from authlib.integrations.flask_client import OAuth
-from functools import wraps
-import json
-from os import environ as env
-
-from six.moves.urllib.request import urlopen
-
-from dotenv import load_dotenv, find_dotenv
-from flask import Flask, jsonify, request, _request_ctx_stack
-from flask_cors import CORS, cross_origin
-from jose import jwt
-
-# Configuration for Auth0 account from .env file
-ENV_FILE = find_dotenv()
-if ENV_FILE:
-    load_dotenv(ENV_FILE)
-AUTH0_DOMAIN = env.get("AUTH0_DOMAIN")
-API_IDENTIFIER = env.get("API_IDENTIFIER")
-ALGORITHMS = ["RS256"]
+from flask import Flask, jsonify, request
+from flask_cors import cross_origin
+from cashman.auth0_decorator import AuthError, requires_auth, requires_scope
 
 app = Flask(__name__)
 
-# Error handler
-class AuthError(Exception):
-    def __init__(self, error, status_code):
-        self.error = error
-        self.status_code = status_code
+# Create sample data
+incomes = [{"description": "salary", "amount": 5000}]
+expenses = [{"description": "pizza", "amount": 50}]
+
+""" 
+Expense could be accessed without authorization --> no token verification needed
+"""
+
+
+@app.route("/expenses/public", methods=["GET"])
+def get_expenses():
+    return jsonify(expenses)
+
+
+@app.route("/expenses/public", methods=["POST"])
+def add_expense():
+    expenses.append(request.get_json())
+    return "", 204
+
+
+@app.route("/expenses/public", methods=["PUT"])
+def replace_expense():
+    expenses.clear()
+    expenses.append(request.get_json())
+    return "", 204
+
+
+@app.route("/expenses/public", methods=["DELETE"])
+def remove_expense():
+    expenses.pop()
+    return "", 204
+
+
+"""
+Income should be accessed with authorization --> auth0 decorator added for each method
+"""
 
 
 @app.errorhandler(AuthError)
@@ -37,126 +51,7 @@ def handle_auth_error(ex):
     return response
 
 
-# Format error response and append status code
-def get_token_auth_header():
-    """Obtains the Access Token from the Authorization Header"""
-    auth = request.headers.get("Authorization", None)
-    if not auth:
-        raise AuthError(
-            {
-                "code": "authorization_header_missing",
-                "description": "Authorization header is expected",
-            },
-            401,
-        )
-
-    parts = auth.split()
-
-    if parts[0].lower() != "bearer":
-        raise AuthError(
-            {
-                "code": "invalid_header",
-                "description": "Authorization header must start with" " Bearer",
-            },
-            401,
-        )
-    elif len(parts) == 1:
-        raise AuthError(
-            {"code": "invalid_header", "description": "Token not found"}, 401
-        )
-    elif len(parts) > 2:
-        raise AuthError(
-            {
-                "code": "invalid_header",
-                "description": "Authorization header must be" " Bearer token",
-            },
-            401,
-        )
-
-    token = parts[1]
-    return token
-
-
-def requires_auth(f):
-    """Determines if the Access Token is valid"""
-
-    @wraps(f)
-    def decorated(*args, **kwargs):
-        token = get_token_auth_header()
-        jsonurl = urlopen("https://" + AUTH0_DOMAIN + "/.well-known/jwks.json")
-        jwks = json.loads(jsonurl.read())
-        unverified_header = jwt.get_unverified_header(token)
-        rsa_key = {}
-        for key in jwks["keys"]:
-            if key["kid"] == unverified_header["kid"]:
-                rsa_key = {
-                    "kty": key["kty"],
-                    "kid": key["kid"],
-                    "use": key["use"],
-                    "n": key["n"],
-                    "e": key["e"],
-                }
-        if rsa_key:
-            try:
-                payload = jwt.decode(
-                    token,
-                    rsa_key,
-                    algorithms=ALGORITHMS,
-                    audience=API_IDENTIFIER,
-                    issuer="https://" + AUTH0_DOMAIN + "/",
-                )
-            except jwt.ExpiredSignatureError:
-                raise AuthError(
-                    {"code": "token_expired", "description": "token is expired"}, 401
-                )
-            except jwt.JWTClaimsError:
-                raise AuthError(
-                    {
-                        "code": "invalid_claims",
-                        "description": "incorrect claims,"
-                        "please check the audience and issuer",
-                    },
-                    401,
-                )
-            except Exception:
-                raise AuthError(
-                    {
-                        "code": "invalid_header",
-                        "description": "Unable to parse authentication" " token.",
-                    },
-                    401,
-                )
-
-            _request_ctx_stack.top.current_user = payload
-            return f(*args, **kwargs)
-        raise AuthError(
-            {"code": "invalid_header", "description": "Unable to find appropriate key"},
-            401,
-        )
-
-    return decorated
-
-
-def requires_scope(required_scope):
-    """Determines if the required scope is present in the Access Token
-    Args:
-        required_scope (str): The scope required to access the resource
-    """
-    token = get_token_auth_header()
-    unverified_claims = jwt.get_unverified_claims(token)
-    if unverified_claims.get("scope"):
-        token_scopes = unverified_claims["scope"].split()
-        for token_scope in token_scopes:
-            if token_scope == required_scope:
-                return True
-    return False
-
-
-incomes = [{"description": "salary", "amount": 5000}]
-expenses = [{"description": "pizza", "amount": 50}]
-
-
-@app.route("/incomes/private")
+@app.route("/incomes/private", methods=["GET"])
 @cross_origin(headers=["Content-Type", "Authorization"])
 @requires_auth
 def get_incomes():
@@ -181,7 +76,7 @@ def replace_income():
 @app.route("/incomes/private", methods=["DELETE"])
 @requires_auth
 def remove_income():
-    # Delete operation requires different scoped permission
+    # Delete operation requires different scope's permission
     if requires_scope("delete:income"):
         incomes.pop()
         return "", 204
@@ -192,27 +87,3 @@ def remove_income():
         },
         403,
     )
-
-
-@app.route("/expenses/public")
-def get_expenses():
-    return jsonify(expenses)
-
-
-@app.route("/expenses/public", methods=["POST"])
-def add_expense():
-    expenses.append(request.get_json())
-    return "", 204
-
-
-@app.route("/expenses/public", methods=["PUT"])
-def replace_expense():
-    expenses.clear()
-    expenses.append(request.get_json())
-    return "", 204
-
-
-@app.route("/expenses/public", methods=["DELETE"])
-def remove_expense():
-    expenses.pop()
-    return "", 204
